Lexical scoping!!
This is a big step forward! (And it wasn't even *that* hard!)
'var' declares a new lexical variable within the current scope.
Scopes are delimited by { }; like Perl, a bare block can be placed as a scope.
Assigning to a variable without 'var' will try to find one in an outer scope if it's not in the current one.
Variables have to be declared, though, to avoid ambiguity.