a new version appears on the horizon!
Rewritten from scratch. Unfortunately, it doesn't have a grammar yet, so it's kind of useless. I don't really want to open that can of worms right now, but I wrote this a while ago and didn't upload it, so I'm doing it now. The new version is located in the apothecary directory, and the old version is located in, shockingly, legacy-version.